requirements/base.txt
anyio
asgiref
attrs
certifi
charset-normalizer
Django
django-cors-headers
django-modeltranslation
django-unfold
djangorestframework
djangorestframework_simplejwt
drf-spectacular
eskiz-sms
h11
httpcore
httpx
idna
inflection
jsonschema
jsonschema-specifications
mysqlclient
phonenumbers
pillow
PyJWT
python-dotenv
python-environ
PyYAML
referencing
requests
rfc3986
rpds-py
sniffio
sqlparse
typing_extensions
uritemplate
urllib3
whitenoise
